refactor : 도메인 함수를 파일 별로 분리
