+++ src/domain/winningResult.js
@@ -1,0 +1,48 @@
+import handleModal from './modal';
+import LottoStatistics from './model/LottoStatistics';
+import { validateBonus, validateLottoNumber } from './validation';
+
+const calculateRevenue = (lottoStatistics, money) => {
+  const revenueRate = lottoStatistics.calculateRevenueRate(
+    lottoStatistics.getProfit(),
+    money,
+  );
+
+  return revenueRate;
+};
+
+const handleLottoStatistics = (lottoStatistics, lottos, winningNumberObj) => {
+  lottoStatistics.compareLottos(lottos, winningNumberObj);
+};
+
+const handleLottoNumberValidate = () => {
+  const winningForm = document.getElementById('winningNumberInputForm');
+  const winningNumbers = Array.from(winningForm.winningNumber).map((input) =>
+    parseInt(input.value, 10),
+  );
+  const bonusNumber = parseInt(winningForm.bonusNumber.value, 10);
+  validateLottoNumber(winningNumbers);
+  validateBonus(bonusNumber, winningNumbers);
+
+  const winningNumberObj = {
+    bonus: bonusNumber,
+    lotto: winningNumbers,
+  };
+  return winningNumberObj;
+};
+
+const handleWinningResult = (e, lottos, money) => {
+  e.preventDefault();
+  try {
+    const winningNumberObj = handleLottoNumberValidate();
+    const lottoStatistics = new LottoStatistics();
+    handleLottoStatistics(lottoStatistics, lottos, winningNumberObj);
+    const revenueRate = calculateRevenue(lottoStatistics, money);
+    const rankResult = lottoStatistics.getRankResult();
+    handleModal(e, rankResult, revenueRate);
+  } catch (error) {
+    alert(error.message);
+  }
+};
+
+export default handleWinningResult;
